fraud-detection-using-vertex-ai/requirement.txt
# Core Libraries
numpy
pandas
scikit-learn
matplotlib
seaborn
tensorflow

# Google Cloud SDKs
google-cloud-aiplatform
google-cloud-storage
google-auth

# Monitoring
alibi-detect

# Utilities
pyyaml
loguru

